backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
